requirements.txt
requests==2.11.1
pymongo==3.11.3
Django==3.2
djangorestframework==3.12.4
pyModbusTCP==0.1.10
dnspython==2.1.0
gunicorn==20.1.0